Rewrite the wavefront engine one last time to remove all the barriers.
This should help mitigate the two-step effect seen in MPI runs, at least to some extent.

The other changes are fixes for bugs that have been there for a while but never noticed
because of the barriers.

Also Sources can now recieve requests for (nearly) anything in (nearly) any order.
Because of HpcrunFSv2 probably won't function, and will require fixes once the proper functions
in prof-lean are reset (if we bother with backwards compat).